GpUnit: refactor created RestClient class - moved some methods from JobRunnerRest into a more general class
